pyproject.toml
[project]
name = "rag-podcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "assemblyai>=0.48.0",
    "beautifulsoup4>=4.14.3",
    "chonkie[qdrant,semantic]>=1.4.2",
    "llama-index>=0.14.10",
    "llama-index-core>=0.14.10",
    "llama-index-vector-stores-qdrant>=0.6.0",
    "llama-index-llms-anthropic>=0.6.0",
    "llama-index-embeddings-voyageai>=0.5.1",
    "lxml>=6.0.2",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.10.0",
    "requests>=2.32.5",
    "rich>=13.0.0",
    "ruff>=0.14.7",
    "sqlalchemy>=2.0.44",
    "tiktoken>=0.12.0",
    "urllib3>=2.5.0",
    "voyageai>=0.3.5",
    "fastmcp>=2.14.1",
    "python-jose[cryptography]>=3.3.0",
    "boto3>=1.42.11",
    "uuid-utils>=0.12.0",
    "psycopg2>=2.9.11",
    "langfuse>=3.11.2",
    "openinference-instrumentation-llama-index>=4.3.9",
    "llama-index-llms-openai>=0.6.10",
    "google-genai>=1.57.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib", "--strict-markers"]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]
